requirements.txt
crewai
streamlit
python-dotenv
chromadb
langchain-openai